Crate tokio_proto

A collection of components for rapid protocol development

Reexports

pub use client::Client;

Modules

client
Utilities for building protocol clients

easy
"Easy" and simplified interfaces to tokio-proto APIs.

multiplex
Dispatch for multiplexed protocols

pipeline
A dispatcher for pipelining protocols

server
A generic Tokio TCP server implementation.

Structs

Body
Body stream

Enums

Error
Error returned as an Error frame or an io::Error that occurerred during normal processing of the Transport

Message
Message sent and received from a multiplexed service

Traits

TryRead
A refinement of std::io::Read for reading from non-blocking sources.

TryWrite
A refinement of std::io::Write for reading from non-blocking sources.