Module tokio_proto::easy

"Easy" and simplified interfaces to tokio-proto APIs.

The tokio-proto pipeline and multiplex modules are relatively complex, namely because of their ability to handle streaming bodies in both requests and responses. Not all protocols, however, need this functionality. To easily get off the ground running with these kinds of protocols, this module contains simplified variants of the pipeline/multiple server/client halves to forgo streaming bodies and instead just deal with requests and responses.

Clients and servers created from the easy module only require an instance of FramedIo, a trait defined in the tokio-core crate for decoding and encoding frames from a byte stream. The frames defined by FramedIo are then the request/response types for the clients and servers defined here.

Note that this module is intended to be paired with tokio_core::easy as well where an instance of EasyFramedIo will slot nicely into all of the implementations here as well.

Internally it's also important to note that all support in this module is built on top of the pipeline and multiplex modules in this crate as well, with most types just taken care of for you.

Modules

multiplex
An "easy" multiplexing module

pipeline
An "easy" pipelining module

Structs

EasyClient
A client which does not support streaming bodies.

EasyResponse
Future returned from Client::call.