Module tokio_proto::easy::pipeline

An "easy" pipelining module

This module is the same as the top-level pipeline module in this crate except that it has no support for streaming bodies. This in turn simplifies a number of generics and allows for protocols to easily get off the ground running.

The API surface area of this module is the same as the top-level one, so no new functionality is introduced here. In fact all APIs here are built on top of the pipeline module itself, they just simplify the generics in play and such.

Structs

EasyServer
An "easy" pipelined server.

Functions

connect
The "easy" form of connecting a pipelined client.