Struct tokio_proto::easy::EasyClient

pub struct EasyClient<R1, R2> {
    // some fields omitted
}

A client which does not support streaming bodies.

This client is returned from the easy::pipeline::connect and easy::multiplex::connect functions in this crate. The client's requests and responses do not support streaming bodies. Each client implements the Service trait to express sending a request to the remote end of this client, and the future returned represents the response. Responses are automatically managed according to the internal protocol.

The R1 type parameter here is the type of requests that this client can send, and is also the argument for the Service::call method trait implementation.

The R2 type parameter is the type of response that will be received and is the resolved value of the future returned by call.

Note that currently this client is not generic over errors as well, but rather all errors returned are instances of io::Error.