Module tokio_proto::pipeline

A dispatcher for pipelining protocols

This module contains reusable components for quickly implementing clients and servers for pipeline based protocols.

Pipelining

Protocol pipelining is a technique in which multiple requests are written out to a single destination without waiting for their corresponding responses. Pipelining is used in a multitude of different protocols such as HTTP/1.1 and Redis in order to increase throughput on a single connection.

Pipelining with the max number of in-flight requests set to 1 implies that for each request, the response must be received before sending another request on the same connection.

Another protocol dispatching strategy is multiplexing (which will be included in Tokio soon).

Usage

Both the server and client pipeline dispatchers take a generic Transport that reads and writes Frame messages. It operates on the transport following the rules of pipelining as described above and exposes the protocol using a Service.

Structs

Pipeline
Provides protocol pipelining functionality in a generic way over clients and servers. Used internally by pipeline::Client and pipeline::Server.

Server
A server Task that dispatches Transport messages to a Service using protocol pipelining.

Enums

Frame
A pipelined protocol frame

Traits

Dispatch
Dispatch messages from the transport to the service

Transport
A specialization of io::Transport supporting the requirements of pipeline based protocols.

Functions

connect
Connect to the given addr and handle using the given Transport and protocol pipelining.

Type Definitions

PipelineMessage
Message used to communicate through the multiplex dispatch