Enum tokio_proto::pipeline::Frame

pub enum Frame<T, B, E> {
    Message(T),
    MessageWithBody(T, Sender<B, E>),
    Body(Option<B>),
    Error(E),
    Done,
}

A pipelined protocol frame

Variants
Message(T)

Either a request or a response

MessageWithBody(T, Sender<B, E>)

Returned by Transport::read when a streaming body will follow. Subsequent body frames will be proxied to the provided Sender.

Calling Transport::write with Frame::MessageWithBody is an error.

Body(Option<B>)

Body frame. None indicates that the body is done streaming.

Error(E)

Error

Done

Final frame sent in each transport direction

Methods

impl<T, B, E> Frame<T, B, E>

fn unwrap_msg(self) -> T

Unwraps a frame, yielding the content of the Message.

fn unwrap_body(self) -> Option<B>

Unwraps a frame, yielding the content of the Body.

fn unwrap_err(self) -> E

Unwraps a frame, yielding the content of the Error.

fn is_done(&self) -> bool

Returns true if the frame is Frame::Done