Module tokio_proto::streaming::multiplex::advanced

Provides the substrate for implementing multiplexed, streaming protocols.

In most cases, it's sufficient to work with streaming::pipeline::{Client, Server} instead. But for some advanced protocols in which the client and servers have more of a peer relationship, it's useful to work directly with these implementation details.

Structs

Multiplex
Task that drives multiplexed protocols

MultiplexMessage
Message used to communicate through the multiplex dispatch

Traits

Dispatch
Dispatch messages from the transport to the service