Trait tokio_proto::streaming::pipeline::Transport

pub trait Transport: 'static + Stream<Error = Error> + Sink<SinkError = Error> {
    fn tick(&mut self) { ... }
    fn cancel(&mut self) -> Result<()> { ... }
}

Additional transport details relevant to streaming, pipelined protocols.

All methods added in this trait have default implementations.

Provided Methods

fn tick(&mut self)

Allow the transport to do miscellaneous work (e.g., sending ping-pong messages) that is not directly connected to sending or receiving frames.

This method should be called every time the task using the transport is executing.

fn cancel(&mut self) -> Result<()>

Cancel interest in the current stream