Trait tokio_proto::BindServer

pub trait BindServer<Kind, T: 'static>: 'static {
    type ServiceRequest;
    type ServiceResponse;
    type ServiceError;
    fn bind_server<S>(&self, handle: &Handle, io: T, service: S) where S: Service<Request = Self::ServiceRequest, Response = Self::ServiceResponse, Error = Self::ServiceError> + 'static;
}

Binds a service to an I/O object.

This trait is not intended to be implemented directly; instead, implement one of the server protocol traits:

pipeline::ServerProto

multiplex::ServerProto

streaming::pipeline::ServerProto

streaming::multiplex::ServerProto

See the crate documentation for more details on those traits.

The Kind parameter, in particular, is a zero-sized type used to allow blanket implementation from the various protocol traits. Any additional implementations of this trait should use their own zero-sized kind type to distinguish them.

Associated Types

type ServiceRequest

The request type for the service.

type ServiceResponse

The response type for the service.

type ServiceError

The error type for the service.

Required Methods

fn bind_server<S>(&self, handle: &Handle, io: T, service: S) where S: Service<Request = Self::ServiceRequest, Response = Self::ServiceResponse, Error = Self::ServiceError> + 'static

Bind the service.

This method should spawn a new task on the given event loop handle which provides the given service on the given I/O object.