Enum tokio_proto::streaming::Message

pub enum Message<T, B> {
    WithoutBody(T),
    WithBody(T, B),
}

Message sent and received from a multiplexed service

Variants

WithoutBody(T)

Has no associated streaming body

WithBody(T, B)

Has associated streaming body

Implementations

impl<T, B> Message<T, B>

pub fn get_ref(&self) -> &T

Returns a reference to the inner value

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the inner value

pub fn into_inner(self) -> T

Consumes the value and returns the inner value

pub fn take_body(&mut self) -> Option<B>

If the Message value has an associated body stream, return it. The original Message value will then become a WithoutBody variant.